feat: decompose CLI into commands
