+++ src/commands/index.ts
@@ -1,0 +1,3 @@
+export { clean } from './clean';
+export { generate } from './generate';
+export { validate } from './validate';
