+++ src/commands/types.ts
@@ -1,0 +1,4 @@
+export type CommmonArgs = {
+  json?: boolean;
+  perf?: boolean;
+};
